task1_sales_chatbot/requirements.txt
fastapi
uvicorn
pandas
pyxlsb
matplotlib
openai
python-dotenv
